added logic for ind_pda table, adjusted schema for fraction inches, added more seed data

+++ 1_drop_and_create_tables.rb
@@ -73,7 +73,7 @@ begin
              flood_insurance varchar(3),
              basement_water varchar(3),
              first_floor_water varchar(3),
-             height_water smallint,
+             height_water numeric(6,3),
              add_comments text,
              assessor_name varchar(50),
              date_assessed date)"
